Improve determination of the number of allowed mismatches

+++ src/main.rs
@@ -84,7 +84,7 @@ fn main() {
         ("map", Some(map_matches)) => {
             let alignment_parameters = utils::AlignmentParameters {
                 base_error_rate: 0.02,
-                poisson_threshold: value_t_or_exit!(map_matches.value_of("poisson_prob"), f32),
+                poisson_threshold: value_t_or_exit!(map_matches.value_of("poisson_prob"), f64),
                 penalty_mismatch: 1,
                 penalty_gap_open: 2,
                 penalty_gap_extend: 1,
